mediafileprevier: Pre discover file synchronously with GES assets
Making the case where we expose the missing asset dialog
working. In the special case we are already blocked on the
main mainloop (as we are handling the missing asset failure)
which means that we can not start another discovery in GstDicoverer.

Discovering should be fast enough for a single asset and we started
doing it async only recently in 48362fe51449132e84d009f54e4da09d25e462dd,

Also make the GES discovering timeout after 5 seconds only, we should not
block the UI for more than that (default is 60secs in GES).

Fixes T7645

``

Differential Revision: https://phabricator.freedesktop.org/D1578