timeline: Fix layers priorities updating when a layer is removed
Until now, when the timeline detected a layer removal, it was updating
the priorities of the rest of the layers itself. This was not good
because when the removal is done when undoing a layer-add operation, the
rest of the layers have their priorities decreased twice: once by the
timeline and once by the LayerMoved actions recorded when the original
undoable action has been performed.

Now the priorities of the remaining layers are decreased at the same
place where the removal is performed: `LayerControls.__delete_layer_cb`.
The Timeline is responsible only for updating the position of the layer
widgets, done through `__update_layers`.

Fixes https://phabricator.freedesktop.org/T7700

Differential Revision: https://phabricator.freedesktop.org/D1660