timeline: Avoid having 2 layers with the same prio in the timeline
This is completely forbiden by GES and should never happen.

GES should not segfault when this case happens but we are anyway
going to replace the layer.priority API with a timeline.move_layer(),
basically because the current API was badly designed and gives to much
constraints on it user (ie. this bug can happen!).

(See https://phabricator.freedesktop.org/D233 for the GES API change
proposal).

Differential Revision: https://phabricator.freedesktop.org/D1674